refactor(server): split export service by domain
- Extract shared excel helpers to shared/utils/excel.util.ts
- Create 5 domain services: document-, task-, deviation-, approval-, user-export
- Rewrite ExportService as thin facade delegating to domain services
- Update ExportModule to register all 5 domain providers
- Rewrite export.service.spec.ts to test facade delegation + DocumentExportService integration
