+++ server/src/modules/export/export.service.spec.ts
@@ -1,11 +1,110 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ExportService } from './export.service';
+import { DocumentExportService } from './services/document-export.service';
+import { TaskExportService } from './services/task-export.service';
+import { DeviationExportService } from './services/deviation-export.service';
+import { ApprovalExportService } from './services/approval-export.service';
+import { UserExportService } from './services/user-export.service';
 import { PrismaService } from '../../prisma/prisma.service';
-import { ExportDocumentsDto, ExportTasksDto, ExportTaskRecordsDto, ExportDeviationReportsDto, ExportApprovalsDto } from './dto';
+import {
+  ExportDocumentsDto,
+  ExportTasksDto,
+  ExportTaskRecordsDto,
+  ExportDeviationReportsDto,
+  ExportApprovalsDto,
+  ExportUsersDto,
+} from './dto';
 
-describe('ExportService', () => {
+describe('ExportService (facade delegation)', () => {
   let service: ExportService;
-  let prisma: any;
+  let documentExport: jest.Mocked<DocumentExportService>;
+  let taskExport: jest.Mocked<TaskExportService>;
+  let deviationExport: jest.Mocked<DeviationExportService>;
+  let approvalExport: jest.Mocked<ApprovalExportService>;
+  let userExport: jest.Mocked<UserExportService>;
+
+  const fakeBuffer = Buffer.from('fake-excel');
+
+  beforeEach(async () => {
+    const mockDocumentExport = { exportDocuments: jest.fn().mockResolvedValue(fakeBuffer) };
+    const mockTaskExport = {
+      exportTasks: jest.fn().mockResolvedValue(fakeBuffer),
+      exportTaskRecords: jest.fn().mockResolvedValue(fakeBuffer),
+    };
+    const mockDeviationExport = { exportDeviationReports: jest.fn().mockResolvedValue(fakeBuffer) };
+    const mockApprovalExport = { exportApprovals: jest.fn().mockResolvedValue(fakeBuffer) };
+    const mockUserExport = { exportUsers: jest.fn().mockResolvedValue(fakeBuffer) };
+
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [
+        ExportService,
+        { provide: DocumentExportService, useValue: mockDocumentExport },
+        { provide: TaskExportService, useValue: mockTaskExport },
+        { provide: DeviationExportService, useValue: mockDeviationExport },
+        { provide: ApprovalExportService, useValue: mockApprovalExport },
+        { provide: UserExportService, useValue: mockUserExport },
+      ],
+    }).compile();
+
+    service = module.get<ExportService>(ExportService);
+    documentExport = module.get(DocumentExportService);
+    taskExport = module.get(TaskExportService);
+    deviationExport = module.get(DeviationExportService);
+    approvalExport = module.get(ApprovalExportService);
+    userExport = module.get(UserExportService);
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('exportDocuments delegates to DocumentExportService', async () => {
+    const dto: ExportDocumentsDto = { level: 1 };
+    const user = { id: 'u1', roleCode: 'admin' };
+    const result = await service.exportDocuments(dto, user);
+    expect(documentExport.exportDocuments).toHaveBeenCalledWith(dto, user);
+    expect(result).toBe(fakeBuffer);
+  });
+
+  it('exportTasks delegates to TaskExportService', async () => {
+    const dto: ExportTasksDto = { status: 'pending' };
+    const result = await service.exportTasks(dto);
+    expect(taskExport.exportTasks).toHaveBeenCalledWith(dto, undefined);
+    expect(result).toBe(fakeBuffer);
+  });
+
+  it('exportTaskRecords delegates to TaskExportService', async () => {
+    const dto: ExportTaskRecordsDto = {};
+    const result = await service.exportTaskRecords(dto);
+    expect(taskExport.exportTaskRecords).toHaveBeenCalledWith(dto, undefined);
+    expect(result).toBe(fakeBuffer);
+  });
+
+  it('exportDeviationReports delegates to DeviationExportService', async () => {
+    const dto: ExportDeviationReportsDto = { status: 'pending' };
+    const result = await service.exportDeviationReports(dto);
+    expect(deviationExport.exportDeviationReports).toHaveBeenCalledWith(dto, undefined);
+    expect(result).toBe(fakeBuffer);
+  });
+
+  it('exportApprovals delegates to ApprovalExportService', async () => {
+    const dto: ExportApprovalsDto = { approverId: 'u2' };
+    const result = await service.exportApprovals(dto);
+    expect(approvalExport.exportApprovals).toHaveBeenCalledWith(dto, undefined);
+    expect(result).toBe(fakeBuffer);
+  });
+
+  it('exportUsers delegates to UserExportService', async () => {
+    const dto: ExportUsersDto = { status: 'active' };
+    const result = await service.exportUsers(dto);
+    expect(userExport.exportUsers).toHaveBeenCalledWith(dto);
+    expect(result).toBe(fakeBuffer);
+  });
+});
+
+describe('DocumentExportService (integration)', () => {
+  let documentExportService: DocumentExportService;
+  let mockPrisma: any;
 
   const mockDocuments = [
     {
@@ -15,526 +114,99 @@ describe('ExportService', () => {
       level: 1,
       version: 1.0,
       status: 'approved',
-      creatorId: 'user-1',
-      approverId: 'user-2',
       createdAt: new Date('2026-01-01'),
       approvedAt: new Date('2026-01-02'),
       creator: { id: 'user-1', name: '张三' },
       approver: { id: 'user-2', name: '李四' },
     },
-    {
-      id: 'doc-2',
-      number: '1-IT-002',
-      title: '需求文档',
-      level: 1,
-      version: 1.2,
-      status: 'pending',
-      creatorId: 'user-3',
-      approverId: null,
-      createdAt: new Date('2026-01-03'),
-      approvedAt: null,
-      creator: { id: 'user-3', name: '王五' },
-      approver: null,
-    },
-  ];
-
-  const mockTasks = [
-    {
-      id: 'task-1',
-      templateId: 'tpl-1',
-      departmentId: 'dept-1',
-      deadline: new Date('2026-02-01'),
-      status: 'pending',
-      creatorId: 'user-1',
-      createdAt: new Date('2026-01-15'),
-      creator: { name: '张三' },
-      department: { name: '技术部' },
-      template: { title: '月度检查' },
-    },
-  ];
-
-  const mockTaskRecords = [
-    {
-      id: 'rec-1',
-      taskId: 'task-1',
-      templateId: 'tpl-1',
-      status: 'completed',
-      submitterId: 'user-1',
-      submittedAt: new Date('2026-02-15'),
-      dataJson: {
-        product_name: '面包',
-        quantity: 1000,
-        production_date: '2026-02-15',
-        approved: true,
-      },
-      submitter: { name: '张三' },
-      template: {
-        id: 'tpl-1',
-        title: '生产记录表',
-        fieldsJson: [
-          { name: 'product_name', label: '产品名称', type: 'text', required: true },
-          { name: 'quantity', label: '生产数量', type: 'number', required: true },
-          { name: 'production_date', label: '生产日期', type: 'date', required: true },
-          { name: 'approved', label: '是否合格', type: 'boolean', required: true },
-        ],
-      },
-    },
-  ];
-
-  const mockDeviationReports = [
-    {
-      id: 'dev-1',
-      recordId: 'rec-1',
-      fieldName: '温度',
-      expectedValue: '100',
-      actualValue: '105',
-      deviationAmount: 5,
-      deviationRate: 0.05,
-      deviationType: 'out_of_range',
-      reason: '设备故障',
-      status: 'pending',
-      reporterId: 'user-1',
-      reportedAt: new Date('2026-01-20'),
-    },
-  ];
-
-  const mockApprovals = [
-    {
-      id: 'app-1',
-      documentId: 'doc-1',
-      approverId: 'user-2',
-      status: 'approved',
-      comment: '通过',
-      createdAt: new Date('2026-01-02'),
-      approvedAt: new Date('2026-01-02'),
-      approver: { name: '李四' },
-      document: { number: '1-IT-001', title: '技术方案' },
-    },
   ];
 
   beforeEach(async () => {
-    const mockPrisma = {
+    mockPrisma = {
       document: {
-        findMany: jest.fn(),
-        count: jest.fn(),
-      },
-      record: {
-        findMany: jest.fn(),
-        count: jest.fn(),
-      },
-      deviationReport: {
-        findMany: jest.fn(),
-        count: jest.fn(),
-      },
-      approval: {
-        findMany: jest.fn(),
-        count: jest.fn(),
-      },
-      user: {
-        findMany: jest.fn(),
-        count: jest.fn(),
+        findMany: jest.fn().mockResolvedValue(mockDocuments),
+        count: jest.fn().mockResolvedValue(1),
       },
     };
 
     const module: TestingModule = await Test.createTestingModule({
       providers: [
-        ExportService,
+        DocumentExportService,
         { provide: PrismaService, useValue: mockPrisma },
       ],
     }).compile();
 
-    service = module.get<ExportService>(ExportService);
-    prisma = module.get(PrismaService);
+    documentExportService = module.get<DocumentExportService>(DocumentExportService);
   });
 
-  afterEach(() => {
-    jest.clearAllMocks();
+  it('should return a buffer with data', async () => {
+    const dto: ExportDocumentsDto = {
+      level: 1,
+      fields: ['number', 'title', 'status'],
+    };
+
+    const buffer = await documentExportService.exportDocuments(dto);
+
+    expect(buffer).toBeInstanceOf(Buffer);
+    expect(buffer.length).toBeGreaterThan(0);
+    expect(mockPrisma.document.count).toHaveBeenCalled();
+    expect(mockPrisma.document.findMany).toHaveBeenCalled();
   });
 
-  describe('exportDocuments', () => {
-    it('应该成功导出文档列表（有数据）', async () => {
-      prisma.document.count.mockResolvedValue(2);
-      prisma.document.findMany.mockResolvedValue(mockDocuments);
-      prisma.user.findMany.mockResolvedValue([
-        { id: 'user-1', name: '张三' },
-        { id: 'user-2', name: '李四' },
-        { id: 'user-3', name: '王五' },
-      ]);
+  it('should return empty workbook buffer when no data', async () => {
+    mockPrisma.document.count.mockResolvedValue(0);
 
-      const dto: ExportDocumentsDto = {
-        level: 1,
-        fields: ['number', 'title', 'level', 'version', 'status', 'creatorName', 'createdAt'],
-      };
+    const dto: ExportDocumentsDto = { level: 2 };
+    const buffer = await documentExportService.exportDocuments(dto);
 
-      const buffer = await service.exportDocuments(dto);
+    expect(buffer).toBeInstanceOf(Buffer);
+    expect(mockPrisma.document.findMany).not.toHaveBeenCalled();
+  });
 
-      expect(buffer).toBeInstanceOf(Buffer);
-      expect(buffer.length).toBeGreaterThan(0);
-      expect(prisma.document.count).toHaveBeenCalled();
-      expect(prisma.document.findMany).toHaveBeenCalled();
-    });
+  it('should apply filters correctly', async () => {
+    mockPrisma.document.count.mockResolvedValue(0);
 
-    it('应该成功导出文档列表（无数据）', async () => {
-      prisma.document.count.mockResolvedValue(0);
+    const dto: ExportDocumentsDto = {
+      level: 1,
+      status: 'approved',
+      keyword: '技术',
+      startDate: '2026-01-01',
+      endDate: '2026-01-31',
+    };
 
-      const dto: ExportDocumentsDto = { level: 2 };
+    await documentExportService.exportDocuments(dto);
 
-      const buffer = await service.exportDocuments(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-      expect(prisma.document.count).toHaveBeenCalledWith({
-        where: { level: 2, deletedAt: null },
-      });
-    });
-
-    it('应该正确应用筛选条件', async () => {
-      prisma.document.count.mockResolvedValue(0);
-
-      const dto: ExportDocumentsDto = {
+    expect(mockPrisma.document.count).toHaveBeenCalledWith({
+      where: expect.objectContaining({
         level: 1,
         status: 'approved',
-        keyword: '技术',
-        startDate: '2026-01-01',
-        endDate: '2026-01-31',
-      };
-
-      await service.exportDocuments(dto);
-
-      expect(prisma.document.count).toHaveBeenCalledWith({
-        where: expect.objectContaining({
-          level: 1,
-          status: 'approved',
-          OR: expect.arrayContaining([
-            { title: { contains: '技术' } },
-            { number: { contains: '技术' } },
-          ]),
-          createdAt: {
-            gte: expect.any(Date),
-            lte: expect.any(Date),
-          },
-          deletedAt: null,
-        }),
-      });
-    });
-
-    it('应该正确处理自定义字段导出', async () => {
-      prisma.document.count.mockResolvedValue(1);
-      prisma.document.findMany.mockResolvedValue([mockDocuments[0]]);
-      prisma.user.findMany.mockResolvedValue([
-        { id: 'user-1', name: '张三' },
-        { id: 'user-2', name: '李四' },
-      ]);
-
-      const dto: ExportDocumentsDto = {
-        fields: ['number', 'title'],
-      };
-
-      const buffer = await service.exportDocuments(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-    });
-
-    it('应该处理大数据量分页导出（性能测试）', async () => {
-      const largeDataCount = 2500;
-      prisma.document.count.mockResolvedValue(largeDataCount);
-
-      // 模拟分页查询
-      prisma.document.findMany.mockImplementation((args: any) => {
-        const skip = args.skip || 0;
-        const take = args.take || 1000;
-        return Promise.resolve(
-          Array(Math.min(take, largeDataCount - skip)).fill(mockDocuments[0])
-        );
-      });
-
-      prisma.user.findMany.mockResolvedValue([
-        { id: 'user-1', name: '张三' },
-        { id: 'user-2', name: '李四' },
-      ]);
-
-      const dto: ExportDocumentsDto = { level: 1 };
-
-      const buffer = await service.exportDocuments(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-      // 应该调用 3 次分页查询 (2500 / 1000 = 3)
-      expect(prisma.document.findMany).toHaveBeenCalledTimes(3);
+        OR: expect.arrayContaining([
+          { title: { contains: '技术' } },
+          { number: { contains: '技术' } },
+        ]),
+        createdAt: {
+          gte: expect.any(Date),
+          lte: expect.any(Date),
+        },
+        deletedAt: null,
+      }),
     });
   });
 
-  describe('exportTasks', () => {
-    it('应该成功导出任务列表', async () => {
-      prisma.record.count.mockResolvedValue(1);
-      prisma.record.findMany.mockResolvedValue(mockTasks);
-
-      const dto: ExportTasksDto = { status: 'pending' };
-
-      const buffer = await service.exportTasks(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-      expect(prisma.record.count).toHaveBeenCalled();
-      expect(prisma.record.findMany).toHaveBeenCalled();
-    });
-
-    it('应该正确应用日期范围筛选', async () => {
-      prisma.record.count.mockResolvedValue(0);
-
-      const dto: ExportTasksDto = {
-        startDate: '2026-01-01',
-        endDate: '2026-01-31',
-      };
-
-      await service.exportTasks(dto);
-
-      expect(prisma.record.count).toHaveBeenCalledWith({
-        where: expect.objectContaining({
-          createdAt: {
-            gte: expect.any(Date),
-            lte: expect.any(Date),
-          },
-          deletedAt: null,
-        }),
-      });
-    });
-  });
-
-  describe('exportDeviationReports', () => {
-    it('应该成功导出偏离报告列表', async () => {
-      prisma.deviationReport.count.mockResolvedValue(1);
-      prisma.deviationReport.findMany.mockResolvedValue(mockDeviationReports);
-
-      const dto: ExportDeviationReportsDto = { status: 'pending' };
-
-      const buffer = await service.exportDeviationReports(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-      expect(prisma.deviationReport.count).toHaveBeenCalled();
-    });
-
-    it('应该正确应用偏离类型筛选', async () => {
-      prisma.deviationReport.count.mockResolvedValue(0);
-
-      const dto: ExportDeviationReportsDto = {
-        deviationType: 'out_of_range',
-      };
-
-      await service.exportDeviationReports(dto);
-
-      expect(prisma.deviationReport.count).toHaveBeenCalledWith({
-        where: expect.objectContaining({
-          deviationType: 'out_of_range',
-          deletedAt: null,
-        }),
-      });
-    });
-  });
-
-  describe('exportApprovals', () => {
-    it('应该成功导出审批记录列表', async () => {
-      prisma.approval.count.mockResolvedValue(1);
-      prisma.approval.findMany.mockResolvedValue(mockApprovals);
-
-      const dto: ExportApprovalsDto = { status: 'approved' };
-
-      const buffer = await service.exportApprovals(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-      expect(prisma.approval.count).toHaveBeenCalled();
-    });
-
-    it('应该正确应用审批人筛选', async () => {
-      prisma.approval.count.mockResolvedValue(0);
-
-      const dto: ExportApprovalsDto = {
-        approverId: 'user-2',
-      };
-
-      await service.exportApprovals(dto);
-
-      expect(prisma.approval.count).toHaveBeenCalledWith({
-        where: expect.objectContaining({
-          approverId: 'user-2',
-        }),
-      });
-    });
-  });
-
-  describe('日期格式化', () => {
-    it('应该正确格式化日期字段', async () => {
-      prisma.document.count.mockResolvedValue(1);
-      prisma.document.findMany.mockResolvedValue([mockDocuments[0]]);
-      prisma.user.findMany.mockResolvedValue([
-        { id: 'user-1', name: '张三' },
-        { id: 'user-2', name: '李四' },
-      ]);
-
-      const dto: ExportDocumentsDto = {
-        fields: ['number', 'createdAt', 'approvedAt'],
-      };
-
-      const buffer = await service.exportDocuments(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-      // Excel 文件应包含格式化的日期
-      const content = buffer.toString();
-      expect(content.length).toBeGreaterThan(0);
-    });
-  });
-
-  describe('边界情况', () => {
-    it('应该处理 null 字段', async () => {
-      prisma.document.count.mockResolvedValue(1);
-      prisma.document.findMany.mockResolvedValue([mockDocuments[1]]);
-      prisma.user.findMany.mockResolvedValue([
-        { id: 'user-3', name: '王五' },
-      ]);
-
-      const dto: ExportDocumentsDto = {
-        fields: ['number', 'approverName', 'approvedAt'],
-      };
-
-      const buffer = await service.exportDocuments(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-    });
-
-    it('应该处理空数组字段', async () => {
-      prisma.document.count.mockResolvedValue(1);
-      prisma.document.findMany.mockResolvedValue([mockDocuments[0]]);
-      prisma.user.findMany.mockResolvedValue([]);
-
-      const dto: ExportDocumentsDto = { fields: [] };
-
-      const buffer = await service.exportDocuments(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-    });
-
-    it('应该处理 undefined 字段', async () => {
-      prisma.document.count.mockResolvedValue(1);
-      prisma.document.findMany.mockResolvedValue([mockDocuments[0]]);
-      prisma.user.findMany.mockResolvedValue([
-        { id: 'user-1', name: '张三' },
-        { id: 'user-2', name: '李四' },
-      ]);
-
-      const dto: ExportDocumentsDto = {};
-
-      const buffer = await service.exportDocuments(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-    });
-  });
-
-  describe('exportTaskRecords', () => {
-    it('应该导出任务记录（包含动态字段）', async () => {
-      prisma.record.count.mockResolvedValue(1);
-      prisma.record.findMany.mockResolvedValue(mockTaskRecords);
-
-      const dto: ExportTaskRecordsDto = {};
-
-      const buffer = await service.exportTaskRecords(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-      expect(buffer.length).toBeGreaterThan(0);
-    });
-
-    it('taskRecordIds 过滤暂未实现，应忽略该字段并以 deletedAt 过滤', async () => {
-      prisma.record.count.mockResolvedValue(1);
-      prisma.record.findMany.mockResolvedValue([mockTaskRecords[0]]);
-
-      const dto: ExportTaskRecordsDto = {
-        taskRecordIds: ['rec-1'],
-      };
-
-      const buffer = await service.exportTaskRecords(dto);
-
-      // Service uses record model with deletedAt filter (taskRecordIds not implemented in where clause)
-      expect(prisma.record.findMany).toHaveBeenCalledWith(
-        expect.objectContaining({
-          where: expect.objectContaining({
-            deletedAt: null,
-          }),
-        }),
+  it('should paginate large datasets', async () => {
+    mockPrisma.document.count.mockResolvedValue(2500);
+    mockPrisma.document.findMany.mockImplementation((args: any) => {
+      const skip = args.skip || 0;
+      const take = args.take || 1000;
+      return Promise.resolve(
+        Array(Math.min(take, 2500 - skip)).fill(mockDocuments[0]),
       );
-      expect(buffer).toBeInstanceOf(Buffer);
     });
 
-    it('taskId 过滤暂未实现，应忽略该字段并以 deletedAt 过滤', async () => {
-      prisma.record.count.mockResolvedValue(1);
-      prisma.record.findMany.mockResolvedValue([mockTaskRecords[0]]);
+    const dto: ExportDocumentsDto = { level: 1 };
+    const buffer = await documentExportService.exportDocuments(dto);
 
-      const dto: ExportTaskRecordsDto = {
-        taskId: 'task-1',
-      };
-
-      const buffer = await service.exportTaskRecords(dto);
-
-      // Service uses record model with deletedAt filter (taskId not implemented in where clause)
-      expect(prisma.record.findMany).toHaveBeenCalledWith(
-        expect.objectContaining({
-          where: expect.objectContaining({
-            deletedAt: null,
-          }),
-        }),
-      );
-      expect(buffer).toBeInstanceOf(Buffer);
-    });
-
-    it('应该按 status 筛选', async () => {
-      prisma.record.count.mockResolvedValue(1);
-      prisma.record.findMany.mockResolvedValue([mockTaskRecords[0]]);
-
-      const dto: ExportTaskRecordsDto = {
-        status: 'completed',
-      };
-
-      const buffer = await service.exportTaskRecords(dto);
-
-      expect(prisma.record.findMany).toHaveBeenCalledWith(
-        expect.objectContaining({
-          where: expect.objectContaining({
-            status: 'completed',
-          }),
-        }),
-      );
-      expect(buffer).toBeInstanceOf(Buffer);
-    });
-
-    it('应该按日期范围筛选', async () => {
-      prisma.record.count.mockResolvedValue(1);
-      prisma.record.findMany.mockResolvedValue([mockTaskRecords[0]]);
-
-      const dto: ExportTaskRecordsDto = {
-        startDate: '2026-02-01',
-        endDate: '2026-02-28',
-      };
-
-      const buffer = await service.exportTaskRecords(dto);
-
-      // Service uses createdAt (not submittedAt) for date range filtering on record model
-      expect(prisma.record.findMany).toHaveBeenCalledWith(
-        expect.objectContaining({
-          where: expect.objectContaining({
-            createdAt: expect.objectContaining({
-              gte: expect.any(Date),
-              lte: expect.any(Date),
-            }),
-          }),
-        }),
-      );
-      expect(buffer).toBeInstanceOf(Buffer);
-    });
-
-    it('应该处理空记录列表', async () => {
-      prisma.record.count.mockResolvedValue(0);
-      prisma.record.findMany.mockResolvedValue([]);
-
-      const dto: ExportTaskRecordsDto = {};
-
-      const buffer = await service.exportTaskRecords(dto);
-
-      expect(buffer).toBeInstanceOf(Buffer);
-    });
+    expect(buffer).toBeInstanceOf(Buffer);
+    expect(mockPrisma.document.findMany).toHaveBeenCalledTimes(3);
   });
 });
